[FIX] account, payment, point_of_sale : prevent deletation of journals
When user is trying to delete the journal we will get a validation error:-
The operation cannot be completed: another model requires the record
being deleted. If possible, archive it instead.

Rather than giving technical errors, we can give proper errors so that user can
easily understand.

Applying these changes will resolve this issue.

Related Enterprise PR - https://github.com/odoo/enterprise/pull/39561

sentry - 4057930920

+++ addons/account/models/account_journal.py
@@ -391,13 +391,23 @@ class AccountJournal(models.Model):
             'alias_name': alias_name,
         }
 
+    def _get_unlink_journals(self):
+        # TO BE OVERRIDE
+        return self.env['account.move'].search([('journal_id', 'in', self.ids)]).mapped('journal_id')
+
     def unlink(self):
         bank_accounts = self.env['res.partner.bank'].browse()
+        journals = self._get_unlink_journals()
         for bank_account in self.mapped('bank_account_id'):
             accounts = self.search([('bank_account_id', '=', bank_account.id)])
             if accounts <= self:
                 bank_accounts += bank_account
         self.mapped('alias_id').sudo().unlink()
+        if journals:
+            raise UserError(
+                _("You cannot delete %s journals because it is link with another model.")
+                % ", ".join(journals.mapped('name'))
+            )
         ret = super(AccountJournal, self).unlink()
         bank_accounts.unlink()
         return ret
